pyproject.toml
[tool.poetry]
name = "compensation-generator"
version = "0.1.0"
description = "Multi-Tenant Commission Calculation System"
authors = ["Martin Sosa <marsosacabrera11@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.1.3"
numpy = "^1.26.2"
pyyaml = "^6.0.1"
boto3 = "^1.29.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
psycopg2-binary = "^2.9.9"
google-auth = "^2.23.4"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.108.0"
gspread = "^5.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
ruff = "^0.8.0"
import-linter = "^2.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short --cov=src --cov-report=term-missing --cov-fail-under=97"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
norecursedirs = [".git", ".tox", "venv", "env", "__pycache__"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/adapters/*",
    "*/infrastructure/repositories/*",
    "*/infrastructure/processing_strategies/*",
    "*/infrastructure/di/bootstrap.py",
    "*/infrastructure/config/*",
    "*/infrastructure/services/*",
    "*/infrastructure/cache/*",
    "*/infrastructure/file_parsers/*",
    "*/infrastructure/validators/*",
    "*/infrastructure/cqrs/*",
    "*/infrastructure/di/container.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
partial_branches = [
    "pragma: no branch",
]
show_missing = true
fail_under = 97

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "C4",
]

[tool.ruff.lint.per-file-ignores]
"src/adapters/cli/main.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
